Break down settngs components
